+++ js/src/jit/mips64/Assembler-mips64.h
@@ -131,7 +131,7 @@ class Assembler : public AssemblerMIPSShared
     static void TraceJumpRelocations(JSTracer* trc, JitCode* code, CompactBufferReader& reader);
     static void TraceDataRelocations(JSTracer* trc, JitCode* code, CompactBufferReader& reader);
 
-    void bind(InstImm* inst, uint64_t branch, uint64_t target);
+    void bind(InstImm* inst, uintptr_t branch, uintptr_t target);
 
     // Copy the assembly code to the given buffer, and perform any pending
     // relocations relying on the target address.
